Rename palette -> frame
