+++ src/frame/contracts.rs
@@ -1,6 +1,6 @@
 //! Implements support for the pallet_contracts module.
 use crate::{
-    palette::{
+    frame::{
         Call,
         balances::Balances,
         system::System,
