+++ src/frame/system.rs
@@ -1,7 +1,7 @@
-//! Implements support for the palette_system module.
+//! Implements support for the frame_system module.
 use crate::{
     error::Error,
-    palette::{
+    frame::{
         Call,
         balances::Balances,
     },
@@ -29,7 +29,7 @@ use runtime_support::Parameter;
 use serde::de::DeserializeOwned;
 use std::fmt::Debug;
 
-/// The subset of the `palette::Trait` that a client must implement.
+/// The subset of the `frame::Trait` that a client must implement.
 pub trait System: 'static + Eq + Clone + Debug {
     /// Account index (aka nonce) type. This stores the number of previous
     /// transactions associated with a sender account.
@@ -80,7 +80,7 @@ pub trait System: 'static + Eq + Clone + Debug {
         + Ord
         + Default;
 
-    /// The address type. This instead of `<palette_system::Trait::Lookup as StaticLookup>::Source`.
+    /// The address type. This instead of `<frame_system::Trait::Lookup as StaticLookup>::Source`.
     type Address: Codec + Clone + PartialEq + Debug;
 
     /// The block header.
@@ -90,9 +90,9 @@ pub trait System: 'static + Eq + Clone + Debug {
 }
 
 /// Blanket impl for using existing runtime types
-impl<T: palette_system::Trait + Debug> System for T
+impl<T: frame_system::Trait + Debug> System for T
 where
-    <T as palette_system::Trait>::Header: serde::de::DeserializeOwned,
+    <T as frame_system::Trait>::Header: serde::de::DeserializeOwned,
 {
     type Index = T::Index;
     type BlockNumber = T::BlockNumber;
