+++ src/runtimes.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with substrate-subxt.  If not, see <http://www.gnu.org/licenses/>.
 
-use crate::palette::{
+use crate::frame::{
     balances::Balances,
     contracts::Contracts,
     system::System,
